# web-app/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.30.0",
    "title": "StudiesPlanner 1 API",
    "description": "Utilize esta documentação para aulixio ao uso da API do sistema!"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": [],
  "produces": [],
  "paths": {
    "/user/": {
      "post": {
        "summary": "Criar usuário",
        "tags": [],
        "description": "Cria um usuário para que seja gerado os valores de key e token para autenticação à API",
        "parameters": [{
          "in": "body",
          "name": "email",
          "required": true,
          "example": "pedro@gmail.com",
          "description": "Email do usuário"
        },{
          "in": "body",
          "name": "password",
          "required": true,
          "example": "Senha123456",
          "description": "Senha do usuário"
        },{
          "in": "body",
          "name": "name",
          "required": false,
          "example": "Pedro",
          "description": "Nome do usuário"
        }],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "Email do usuário cadastrado."
                },
                "x-key-api": {
                  "type": "string",
                  "example": "Key de autenticação do usuário para acessar a API."
                },
                "x-token-api": {
                  "type": "string",
                  "example": "Token de autenticação do usuário para acessar a API."
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar um email de usuário e senha para cadastro!"
                }
              }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "put": {
        "tags": [],
        "summary": "Atualizar usuário",
        "description": "Atualiza o nome do usuário",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },{
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          },
          {
            "in": "body",
            "name": "name",
            "required": false,
            "example": "Pedro Henrique",
            "description": "Nome do usuário"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Usuário atualizado!"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Não é possível atualizar a senha por essa rota de API!"
                }
              }
            }
          }
        }
      }
    },
...
    "/task/": {
      "get": {
        "tags": [],
        "summary": "Retorna todas as tarefas",
        "description": "Retorna todas as tarefas de todas as disciplinas cadastradas para o usuário",
        "parameters": [
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "tasks": {
                  "type": "array",
                  "example":
                  [{
                    "name": "Nome de identificação da tarefa",
                    "subject": {
                      "id": "IDDISCIPLINA",
                      "author": "IDUSUARIO"
                    },
                    "observation": "Lista de revisão - Se tiver dúvidas perguntar depois",
                    "deadline": "2020-07-28",
                    "realizedDate": "2020-07-29",
                    "realized": "true",
                    "create": "2020-06-10"
                  }]
                }
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "Criar tarefa",
        "description": "Cria uma tarefa para uma disciplina em específico",
        "parameters": [
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          },
          {
            "in": "body",
            "name": "name",
            "required": true,
            "example": "Lista Um",
            "description": "Nome de identificação da tarefa"
          },
          {
            "in": "body",
            "name": "subject",
            "required": true,
            "example": "abcd1234",
            "description": "Id da disciplina em que a tarefa pertence"
          },
          {
            "in": "body",
            "name": "deadline",
            "required": true,
            "example": "2020-08-22",
            "description": "Data de entrega da tarefa"
          },
          {
            "in": "body",
            "name": "observation",
            "required": false,
            "example": "Se houver dúvia mandar por email",
            "description": "Uma observação pertinente à tarefa"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Tarefa criada!"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar uma tarefa para ser salva!"
                }
              }
            }
          }
        }
      }
    },
    "/task/{id}": {
      "get": {
        "tags": [],
        "summary": "Retorna uma tarefa em específico",
        "description": "Retorna a tarefa correspondente ao id passado",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Lista Um"
                },
                "observation": {
                  "type": "string",
                  "example": "Se houver dúvia mandar por email"
                },
                "deadline": {
                  "type": "date",
                  "example": "2020-08-26"
                },
                "subject": {
                  "type": "object",
                  "example": {
                    "id": "IDDISCIPLINA",
                    "author": "IDUSUARIO"
                  }
                },
                "realized": {
                  "type": "boolean",
                  "example": true
                },
                "realizedDate": {
                  "type": "date",
                  "example": "2020-08-22"
                },
                "create": {
                  "type": "date",
                  "example": "2020-08-26"
                }
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [],
        "summary": "Atualizar tarefa",
        "description": "Permite atualizar o nome, observação e data de entrega da tarefa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          },
          {
            "in": "body",
            "name": "name",
            "required": false,
            "example": "Lista Um",
            "description": "Nome de identificação da tarefa"
          },
          {
            "in": "body",
            "name": "deadline",
            "required": false,
            "example": "2020-08-22",
            "description": "Data de entrega da tarefa"
          },
          {
            "in": "body",
            "name": "observation",
            "required": false,
            "example": "Se houver dúvia mandar por email",
            "description": "Uma observação pertinente à tarefa"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Tarefa atualizada!"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Não é possível atualizar o author da tarefa!"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [],
        "summary": "Deletar tarefa",
        "description": "Deleta uma tarefa específica do usuário",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Tarefa deletada!"
                }
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                }
              }
            }
          }
        }
      }
    },
    "/task/{id}/realized": {
      "put": {
        "tags": [],
        "summary": "Registrar tarefa como entregue",
        "description": "Registra a tarefa como entregue, marcando assim a data com que a tarefa foi entregue",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Tarefa marcada como realizada!"
                }
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "parameters": {
    "x-key-api": {
      "name": "x-key-api",
      "in": "header",
      "description": "API Key do seu usuário",
      "required": true,
      "type": "string"
    },
    "x-token-api": {
      "name": "x-token-api",
      "in": "header",
      "description": "API Key do seu usuário",
      "required": true,
      "type": "string"
    }
  }
}